Hanshuai/issue feature001

+++ src/MaterialCodeSelectionPlatform/MaterialCodeSelectionPlatform.Domain/DTO/MaterialTakeOffDetailDto.cs
@@ -20,7 +20,7 @@
         public string MaterialTakeOffDetailId { get; set; }
 
 
-        public double? RoundUp { get; set; }
+        public float? RoundUp { get; set; }
 
         public string PartNumber { get; set; }
     }
